+++ Models/Chat/DirectChat.cs
@@ -11,6 +11,6 @@ namespace ServicesLibrary.Models.Chat
         [JsonProperty("id")] public BigInteger Id { get; set; }
         [JsonProperty("members")] public List<User> Members { get; set; }
         [JsonProperty("chat_type")] public string ChatType { get; set; } = TypesOfChat.DirectChat;
-        [JsonProperty("updated_at")] public int UpdatedAt { get; set; }
+        [JsonProperty("updated_at")] public long UpdatedAt { get; set; }
     }
 }
