+++ Models/Chat/Group.cs
@@ -16,6 +16,6 @@ namespace ServicesLibrary.Models.Chat
         [JsonProperty("chat_type")] public string ChatType { get; set; } = TypesOfChat.Group;
         [JsonProperty("photo_url")] public string PhotoUrl { get; set; }
         [JsonProperty("members_count")] public int MembersCount { get; set; }
-        [JsonProperty("updated_at")] public int UpdatedAt { get; set; }
+        [JsonProperty("updated_at")] public long UpdatedAt { get; set; }
     }
 }
